scripts/knnSearches/requirements.txt
Package                      Version
---------------------------- ----------------
torch                        1.13.0
elasticsearch                8.8.2
pandas                       1.5.3
PyYAML                       5.4.1